+++ host-monitoring-worker/frontend/chunks/app-KVGSCKNB.js
@@ -27,7 +27,7 @@ import {
   useRef,
   useState,
   useSyncExternalStore
-} from "./chunk-PLSZNCEC.js";
+} from "./chunk-3IENDW2S.js";
 
 // src/jsx-runtime.ts
 function jsx(type, props, key) {